Building vote main site
