+++ frontend/package.json
@@ -19,6 +19,7 @@
     "react-bootstrap": "^1.4.0",
     "react-csv": "^2.0.3",
     "react-dom": "^17.0.1",
+    "react-icons": "^4.1.0",
     "react-perfect-scrollbar": "^1.5.8",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
