+++ frontend/src/App.js
@@ -18,6 +18,7 @@ import RegisterEvent from "./components/RegisterEvent";
 import GuideVoters from "./components/GuideVoters";
 import GuideParticipants from "./components/GuideParticipants";
 import Vote from "./components/Vote";
+import VoteEnvSemi from "./components/VoteEnvSemi";
 
 class App extends Component {
   constructor(props) {
@@ -77,6 +78,7 @@ class App extends Component {
             />
             <Route exact path="/guide/voters" component={GuideVoters} />
             <Route exact path="/vote" component={Vote} />
+            <Route exact path="/vote/semi/env" component={VoteEnvSemi} />
             <Redirect from="/" to="/" />
           </Switch>
         </BrowserRouter>
