+++ frontend/src/components/VotePage.css
@@ -29,7 +29,6 @@
   align-items: center;
   justify-content: center;
   text-align: center;
-  line-height: 30px;
 }
 
 .main-btn-join {
